test: move dbfixtures and testmode to storage/internal.
